plot/24-hour/requirements.txt
requests==2.25.1
matplotlib==3.3.4
numpy==1.20.0
python-dotenv==0.15.0
